Implemented page views, updated cmd and statistics
Implemented page views to support extracting information from multiple
pages, updated the statistics calculator and cmd to operate on those
page views. Added the ability to set file name so that the pickle file
was not always saved as filename.pickle, user has to think about the
filename now when loading the file, they have the option to set it
before loading, the filename has a default though should the user not
care about file names.