Duplicate Code Refactoring - see description
Step 1 - A function has already been written with needed functionality
due to previous refactoring step. However, that step was refactoring
long method, and as a result, did not cater to duplicated code (code
that this function was handling) in other areas.
Step 2 - This commit is catering to the next step which involves
removing duplicate code and replacing with function call